requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
configobj==5.0.8
configparser==5.3.0
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.12.0
fitz==0.0.1.dev2
Flask==2.3.2
Flask-Login==0.6.2
fpdf==1.7.2
future==0.18.3
httplib2==0.22.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
looseversion==1.1.2
lxml==4.9.2
MarkupSafe==2.1.2
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pathlib==1.0.1
Pillow==9.5.0
prov==2.0.0
psycopg2==2.9.6
pydot==1.4.2
PyMuPDF==1.22.3
pyparsing==3.0.9
PyPDF2==3.0.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyxnat==1.5
qrcode==7.4.2
rdflib==6.3.2
requests==2.30.0
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
traits==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
ua-parser==0.16.1
urllib3==2.0.2
user-agents==2.2.0
Werkzeug==2.3.4
